Create PlatformFxRuntime
